pyproject.toml
[project]
name = "voice-record-translate"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pywebview>=5.0",
    "flask>=3.0",
    "flask-cors>=4.0",
    "openai==2.23.0",
    "av>=12.0",
    "soniox",
]

[project.scripts]
vrt = "vrt.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
